+++ test/integration/api_uninitialized_test.lua
@@ -97,11 +97,11 @@ function g.test_uninitialized()
 
     local resp = g.server:graphql({
         query = [[{
-            cluster { failover }
+            cluster { failover {enabled} }
         }]]
     })
 
-    t.assert_equals(resp['data']['cluster']['failover'], false)
+    t.assert_equals(resp.data.cluster.failover.enabled, false)
 
     t.assert_error_msg_contains(
         "Cluster isn't bootstrapped yet",
@@ -109,7 +109,7 @@ function g.test_uninitialized()
             return g.server:graphql({
                 query = [[
                     mutation {
-                        cluster { failover(enabled: false) }
+                        cluster { failover(enabled: false) {} }
                     }
                 ]]
             })